build: fixs on vercel
