+++ .gitignore
@@ -14,3 +14,5 @@ apps/web/public/assets/img/
 packages/core/src/constants/data.mjs
 
 .vercel
+
+.vercel
