+++ .vercelignore
@@ -10,3 +10,11 @@ scripts
 CHANGELOG.md
 README.md
 
+*.log
+.output
+.nuxt
+.idea
+.vercel
+eslint.config.ts
+LICENSE
+.editorconfig
