+++ apps/monkey/package.json
@@ -14,6 +14,11 @@
   },
   "devDependencies": {
     "@types/file-saver": "^2.0.7",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "unplugin-auto-import": "^0.17.5",
+    "unplugin-vue-components": "^0.26.0",
+    "unplugin-vue-macros": "^2.7.10",
+    "vite": "^5.1.4",
     "vite-plugin-monkey": "^3.5.1"
   }
 }
