+++ apps/monkey/vite.config.ts
@@ -1,17 +1,46 @@
+import path from 'node:path'
 import { defineConfig } from 'vite'
 import monkey, { cdn, util } from 'vite-plugin-monkey'
 import AutoImport from 'unplugin-auto-import/vite'
+import Vue from '@vitejs/plugin-vue'
+import vueJsx from '@vitejs/plugin-vue-jsx'
+import UnoCSS from 'unocss/vite'
 
-import config, { autoImportConfig } from '../../vite.config'
+export const root = path.resolve(__dirname, '../../')
+export const packages = path.resolve(root, 'packages')
+export const core = path.resolve(packages, 'core/src')
+const ui = path.resolve(packages, 'ui/src')
 
 const repo = 'https://github.com/Chilfish/Weibo-archiver'
 const downloadURL = `${repo}/raw/monkey/weibo-archiver.user.js`
 
 export default defineConfig({
-  ...config,
+  resolve: {
+    alias: {
+      '@core': core,
+      '@ui': ui,
+    },
+  },
+  build: {
+    minify: true,
+    outDir: path.join(root, 'dist/monkey'),
+  },
   plugins: [
-    ...config.plugins!,
-    AutoImport(autoImportConfig),
+    Vue(),
+    UnoCSS(),
+    vueJsx(),
+    AutoImport({
+      imports: [
+        'vue',
+        '@vueuse/core',
+      ],
+      dts: path.resolve(root, 'types/auto-imports.d.ts'),
+      dirs: [
+        core,
+        ui,
+      ],
+      vueTemplate: true,
+    }),
 
     monkey({
       entry: 'src/main.ts',
@@ -48,6 +77,9 @@ export default defineConfig({
           ],
         },
         fileName: 'weibo-archiver.user.js',
+        externalResource: {
+          'naive-ui/dist/index.css': cdn.unpkg(),
+        },
       },
     }),
   ],
