+++ package.json
@@ -34,15 +34,10 @@
     "@vueuse/core": "^10.8.0",
     "destr": "^2.0.3",
     "pinia": "^2.1.7",
-    "vue": "^3.4.19",
-    "vue-router": "^4.3.0"
+    "vue": "^3.4.19"
   },
   "devDependencies": {
     "@antfu/eslint-config": "^2.6.4",
-    "@types/node": "^20.11.20",
-    "@vitejs/plugin-vue": "^5.0.4",
-    "@vitejs/plugin-vue-jsx": "^3.1.0",
-    "@vue-macros/volar": "^0.18.11",
     "eslint": "npm:eslint-ts-patch@8.55.0-1",
     "eslint-plugin-format": "^0.1.0",
     "eslint-ts-patch": "8.55.0-1",
@@ -51,12 +46,6 @@
     "naive-ui": "^2.38.0",
     "pnpm": "^8.15.3",
     "typescript": "^5.3.3",
-    "unplugin-auto-import": "^0.17.5",
-    "unplugin-vue-components": "^0.26.0",
-    "unplugin-vue-macros": "^2.7.10",
-    "vite": "^5.1.4",
-    "vite-plugin-inspect": "^0.8.3",
-    "vite-plugin-vue-devtools": "1.0.0-rc.8",
     "vue-tsc": "^1.8.27"
   },
   "lint-staged": {
