+++ packages/core/src/utils/index.ts
@@ -30,10 +30,10 @@ export function delay(ms = 2000) {
 }
 
 export async function getOptions() {
-  if (typeof localStorage !== 'undefined') {
+  if (!isElectron) {
     return JSON.parse(localStorage.getItem('fetchOptions') || '{}') as FetchOptions
   }
-  // TODO: 在构建 monkey, web 时，注释下面几行
+  // TODO: 在构建 monkey 时，注释下面几行
   else {
     const { config } = await import('./config')
     return config.store.fetchOptions
