+++ pnpm-workspace.yaml
@@ -1,3 +1,3 @@
 packages:
-  - 'packages/**'
-  - 'apps/**'
+  - 'packages/*'
+  - 'apps/*'
