+++ types/auto-components.d.ts
@@ -12,19 +12,15 @@ declare module 'vue' {
     ImgViewer: typeof import('./../packages/ui/src/ImgViewer.vue')['default']
     MainHeader: typeof import('./../packages/ui/src/MainHeader.vue')['default']
     MainImage: typeof import('./../packages/ui/src/MainImage.vue')['default']
-    NAvatar: typeof import('naive-ui')['NAvatar']
     NButton: typeof import('naive-ui')['NButton']
     NForm: typeof import('naive-ui')['NForm']
     NFormItem: typeof import('naive-ui')['NFormItem']
     NImage: typeof import('naive-ui')['NImage']
-    NImageGroup: typeof import('naive-ui')['NImageGroup']
     NInput: typeof import('naive-ui')['NInput']
     NModalProvider: typeof import('naive-ui')['NModalProvider']
-    NPagination: typeof import('naive-ui')['NPagination']
     NPopconfirm: typeof import('naive-ui')['NPopconfirm']
     NRadio: typeof import('naive-ui')['NRadio']
     NRadioGroup: typeof import('naive-ui')['NRadioGroup']
-    NSpin: typeof import('naive-ui')['NSpin']
     NSwitch: typeof import('naive-ui')['NSwitch']
     NTabPane: typeof import('naive-ui')['NTabPane']
     NTabs: typeof import('naive-ui')['NTabs']
